+++ example/oglplus/026_clouds.cpp
@@ -4,7 +4,7 @@
  *
  *  @oglplus_screenshot{026_clouds}
  *
- *  Copyright 2008-2014 Matus Chochlik. Distributed under the Boost
+ *  Copyright 2008-2015 Matus Chochlik. Distributed under the Boost
  *  Software License, Version 1.0. (See accompanying file
  *  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
  *
@@ -277,7 +277,7 @@ public:
 		gl.Viewport(width, height);
 		Mat4f perspective = CamMatrixf::PerspectiveX(
 			Degrees(65),
-			double(width)/height,
+			float(width)/height,
 			1, 40
 		);
 		ProgramUniform<Mat4f>(cloud_prog, "ProjectionMatrix").Set(perspective);
