+++ example/oglplus/027_brain_scan_vis.cpp
@@ -9,7 +9,7 @@
  *  from a medical scan stored in MINC format using the *nix minc-tools
  *  (mincinfo, mincreshape and minctoraw).
  *
- *  Copyright 2008-2014 Matus Chochlik. Distributed under the Boost
+ *  Copyright 2008-2015 Matus Chochlik. Distributed under the Boost
  *  Software License, Version 1.0. (See accompanying file
  *  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
  *
@@ -326,7 +326,7 @@ public:
 		prog.projection_matrix.Set(
 			CamMatrixf::PerspectiveX(
 				Degrees(25),
-				double(width)/height,
+				float(width)/height,
 				1, 20
 			)
 		);
